Abstract gross-payment-chart into main file and core file
